+++ src/lib/libdwarf/dwarf_memcpy_swap.h
@@ -33,12 +33,12 @@ extern "C" {
 #endif /* __cplusplus */
 
 void _dwarf_memcpy_swap_bytes(void *s1, const void *s2,
-    size_t len);
+    unsigned long len);
 /*  It was once inconvenient to use memcpy directly as it
     uses size_t and that requires <stddef.h>,
     although stddef.h is a part of C90, so..ok. */
 void _dwarf_memcpy_noswap_bytes(void *s1,
-    const void *s2, size_t len);
+    const void *s2, unsigned long len);
 
 #ifdef __cplusplus
 }
